Correct handling of member names in LongLink blocks.
Case 74517: Archive::Tar::Builder's "LongLink" compatibility
code neglected to preserve the behavior of archive_as, which
allows you to specify a member name for the file as archived
in the tarball. Instead, when the file length limit was hit
and LongLink mode kicked in (when the length exceeded 155),
it fell back to archiving with the origin path name.

This commit adjusts the LongLink code to use the member name
rather than the origin path when writing the LongLink blocks.

Note that this is not the same problem as case 71693, but the
symptom is similar. In both cases, what you end up seeing is
certain paths represented as absolute (e.g. /home/user/...)
rather than relative in the archive.