Do not add a trailing slash to names of files.
Case 74781: The header code was appending a slash to the prefix portion of a
filename even if there was no suffix.  This resulted in filenames with a length
of between 100 and 156 characters having a trailing slash, even if they did not
represent directories.  This caused tar to unpack them as directories and exit
unsuccessfully.  Fix this by not adding this slash if there is no suffix.