Fixed support for file sizes between 4 & 8 GiB
Case 80933: Corrected the code-path for files <8 GiB by fixing the printf
statement so that it knows the filesize value is a 64-bit integer, instead
of interpreting it as a 32-bit one.  Now files between [4 GiB, 8 GiB) will
correctly be added to the created tarball.

Furthermore, fixed the border case where files exactly 8 GiB in size would
use the <8 GiB tar format instead of the >=8 GiB (extended) format.

Additionally, the bigfile test case was updated to test more (and larger)
big files to avoid future issues.