open(), fstat() files before writing headers
It appears that in some cases there is a disconnect between when the
file is stat'd and when it is opened, resulting in a race window where
the file can disappear out from under the archiver.  This produces an
error message, and previously one that was misleading, as it referred to
the wrong file.

This patch is a modified version of one upstream applied to fix a
different problem, but it will likely fix this race condition as well.
Furthermore, it produces a more useful error message, which is desirable
in and of itself.  The original commit message follows.

open() and fstat() files before writing headers, to ensure size
indicated is the same as size of file archived; the following technique
is used to achieve this:

    1. open() each inode with O_NOFOLLOW, indiscriminately.  If open()
       returns -1 and sets errno to ELOOP, then use lstat() to obtain
       symlink inode information.

    2. If open() succeeds, use fstat() subsequently, keeping the fd open
       for the duration of the file archival.

This patch also opens files with O_NONBLOCK in order to ensure
that that reading a pipe does not cause a hang.