Add support for long symlink destinations.
If a symlink destination is longer than 100 characters, it won't fit
into the tar header.  Archive::Tar::Builder used to just truncate these
entries.  Instead, have it encode them using either a GNU LongLink
header (type 'K') or a POSIX linkpath header record, as appropriate.

Note that GNU tar doesn't like it if you use a LongLink encoding when it
isn't required (i.e., for entries that fit normally in the header), so
we don't do that.  Making this work properly requires more code than the
lazy way.

Add a test to ensure that we encode the symlink destination properly as
well as the symlink's name itself.  This requires that we use tar -tvf,
but that appears to do the right thing on both GNU tar and BSD tar.
star appears to be broken in that case, but it was already broken in
other ways.