Fix corrupt pax-style archives on 32-bit systems.
When performing an snprintf to encode the size of a pax extended header,
we were passing a size_t (a perfectly reasonable type for header size)
for a format expecting an unsigned long long.  As a result, we encoded
four extra bytes of junk in the high bits on little-endian 32-bit
systems, and probably produced equally corrupt data on big-endian
systems.  This led to the generated archive being rejected for having an
absurdly large size for the pax extended header.  Cast the value to
unsigned long long so the compiler knows how to pass it through to
snprintf.